### 2022-10-27 #104 @whedon
[PRE REVIEW]: AstroImages.jl
**Submitting author:** @sefffal (<a href="http://orcid.org/0000-0001-5684-4593">William R. Thompson</a>)
**Repository:** <a href="https://github.com/JuliaAstro/AstroImages.jl.git" target ="_blank">https://github.com/JuliaAstro/AstroImages.jl.git</a>
**Version:**
