### 2022-10-27 #104 @sefffal
Not sure what the issue is. The paper is in the expected location paper.tex in the paper folder on the paper branch of the repo.
