### 2022-11-01 ae707d6
Use SnoopPrecompile and Fix precompile breakage on nightly (#35)
* Use SnoopPrecompile

* Correct compat bounds on SnoopPrecompile

* Add missing newline

Co-authored-by: Mosè Giordano <giordano@users.noreply.github.com>
