+++ Project.toml
@@ -18,6 +18,7 @@ MappedArrays = "dbb5928d-eab1-5f90-85c2-b9b0edb7c900"
 PlotUtils = "995b91a9-d308-5afd-9ec6-746e21dbc043"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
+SnoopPrecompile = "66db9d55-30c0-4569-8b51-7e840670fc0c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
@@ -37,6 +38,7 @@ ImageShow = "0.3"
 MappedArrays = "0.4"
 PlotUtils = "1.2"
 RecipesBase = "1.2"
+SnoopPrecompile = "1"
 Tables = "1.7"
 WCS = "0.6"
 julia = "^1.6.0"
