### 2022-11-02 #104 @sefffal
The issue seems to be that the `paper` directory is not merged on the main branch. From the submission instructions, it seemed it possible to keep it on a separate `paper` branch. Is that still ok, or would I have to merge it?
