### 2022-11-04 086cbde
CompatHelper: bump compat for DimensionalData to 0.23, (keep existing compat)
