### 2022-11-07 47c3cc7
CompatHelper: bump compat for FITSIO to 0.17, (keep existing compat) (#37)
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
