+++ Project.toml
@@ -29,7 +29,7 @@ AbstractFFTs = "1.1"
 AstroAngles = "0.1"
 ColorSchemes = "3.18"
 DimensionalData = "0.20.8"
-FITSIO = "0.16"
+FITSIO = "0.16, 0.17"
 FileIO = "1.15"
 ImageAxes = "0.6"
 ImageBase = "^0.1.5"
