+++ Project.toml
@@ -27,7 +27,7 @@ WCS = "15f3aee2-9e10-537f-b834-a6fb8bdb944d"
 AbstractFFTs = "1.1"
 AstroAngles = "0.1"
 ColorSchemes = "3.18"
-DimensionalData = "0.22"
+DimensionalData = "0.23"
 FITSIO = "0.16"
 FileIO = "1.15"
 ImageAxes = "0.6"
