### 2022-11-09 cb9feca
Fix dims after imview. * header now shared more often when taking views of data * restored imview live update functionality (actually behave like a view) * Removed ability to call imview then implot and keep axes. This was never going to work in the long run.
