### 2022-11-10 26f8189
Bump DimensionalData to fix broadcasting  (#34)
* Bump compat with DimensionalData

* Fix ambiguity error with newer DimensionalData

* Skip right to DimensionalData 0.23

* Fix dims after imview.
* header now shared more often when taking views of data
* restored imview live update functionality (actually behave like a view)
* Removed ability to call imview then implot and keep axes. This was never going to work in the long run.

* Fix tests

* Align composecolors with new behavior of imview
