### 2022-11-10 be5868b
Use `load` instead of `AstroImage` in docs.
