+++ docs/Project.toml
@@ -1,17 +1,17 @@
 [deps]
-AstroImages = "fe3fc30c-9b16-11e9-1c73-17dabf39f4ad"
-Contour = "d38c429a-6771-53c6-b99e-75d170b6e991"
-DimensionalData = "0703355e-b756-11e9-17c0-8b28908087d0"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
-FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
-ImageFiltering = "6a3955dd-da59-5b1f-98d4-e7296123deb5"
-ImageIO = "82e4d734-157c-48bb-816b-45c225c6df19"
-ImageTransformations = "02fcd773-0e25-5acc-982a-7f6622650795"
-Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
-Photometry = "af68cb61-81ac-52ed-8703-edc140936be4"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Reproject = "d1dcc2e6-806e-11e9-2897-3f99785db2ae"
 WCS = "15f3aee2-9e10-537f-b834-a6fb8bdb944d"
+ImageFiltering = "6a3955dd-da59-5b1f-98d4-e7296123deb5"
+Photometry = "af68cb61-81ac-52ed-8703-edc140936be4"
+DimensionalData = "0703355e-b756-11e9-17c0-8b28908087d0"
+ImageIO = "82e4d734-157c-48bb-816b-45c225c6df19"
+Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+AstroImages = "fe3fc30c-9b16-11e9-1c73-17dabf39f4ad"
+FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
+Contour = "d38c429a-6771-53c6-b99e-75d170b6e991"
+ImageTransformations = "02fcd773-0e25-5acc-982a-7f6622650795"
 
 [compat]
 Documenter = "0.27"
