### 2022-12-12 #104 @sefffal
I’m not sure what would cause this issue. I placed the template in the paper directory on the paper branch of the repo.
Any suggestions?
