### 2022-12-12 #104 @carstenbauer
There is this one ☝️ but it might not be very up to date, so feel free to suggest other people as well (can also be outside the Julia community if you think they might still be willing to review).
