### 2022-12-16 0b49955
CompatHelper: bump compat for DimensionalData to 0.24, (keep existing compat)
