+++ previews/PR39/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"guide/contours/#Contours","page":"Plotting Contours","title":"Contours","text":"","category":"section"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"This guide shows a few different ways to measure and visualize contours of images.","category":"page"},{"location":"guide/contours/#Using-Plots","page":"Plotting Contours","title":"Using Plots","text":"","category":"section"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"The most basic way to create a contour plot is simply to use Plots.jl contour and contourf functions on your image.","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"Let's see how that works:","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"using AstroImages, Plots\n\n\n# First load a FITS file of interest\nfname = download(\n    \"http://www.astro.uvic.ca/~wthompson/astroimages/fits/herca/herca_radio.fits\",\n    \"herca-radio.fits\"\n)\n\nherca = load(\"herca-radio.fits\")","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"Create a contour plot","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"contour(herca)","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"Create a filled contour plot","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"contourf(herca)","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"Specify the number of levels","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"contour(herca, levels=5)","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"Specify specific levels","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"contour(herca, levels=[1, 1000, 5000])","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"Overplot contours on image:","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"implot(herca)\ncontour!(herca, levels=4, color=:cyan)","category":"page"},{"location":"guide/contours/#Using-Contour.jl","page":"Plotting Contours","title":"Using Contour.jl","text":"","category":"section"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"For more control over how contours are calculated and plotted, you can use the Contour.jl package:","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"using Contour\nherca = load(\"herca-radio.fits\")\n\np = implot(herca, cmap=nothing, colorbar=false)\n\n# Note: Contour.jl only supports float inputs https://github.com/JuliaGeometry/Contour.jl/issues/73\nfor cl in levels(contours(dims(herca)..., float.(herca)))\n    lvl = level(cl) # the z-value of this contour level\n    for line in lines(cl)\n        xs, ys = coordinates(line) # coordinates of this line segment\n        plot!(p, xs, ys, line_z=lvl, label=\"\")\n    end\nend\n\np","category":"page"},{"location":"guide/contours/","page":"Plotting Contours","title":"Plotting Contours","text":"Here we plot just the contours, now i
[…]
