### 2022-12-28 #104 @carstenbauer
@sefffal Quick reminder of the TODOs:

1) fix the `.latexmkrc` issue
2) suggest reviewers as soon as possible

Enjoy the rest of the holidays / the turn of the year!
