### 2022-12-28 #104 @sefffal
Thanks for the reminder @carstenbauer .
I've added the missing latexmk file (sorry about that!). Maybe there is another issue.

Here are some suggested reviewers: lucaferranti, mauro3, bidhya, hameye
