### 2023-01-11 #104 @sefffal
Appologies but I might need some help debugging this. My paper compiles locally using latexmk, and I'm following the template closely. For example, the `bezanson2017julia` citation is taken verbatim from https://github.com/JuliaCon/JuliaConSubmission.jl/

In the meantime I committed my locally compiled PDF to show what it should look like:
https://github.com/JuliaAstro/AstroImages.jl/blob/paper/paper/paper.pdf
