### 2023-02-17 7c3c218
CompatHelper: bump compat for DimensionalData to 0.24, (keep existing compat) (#39)
* CompatHelper: bump compat for DimensionalData to 0.24, (keep existing compat)

* Update Project.toml

---------

Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
Co-authored-by: Will Thompson <will.thompson@outlook.com>
