+++ Project.toml
@@ -1,7 +1,7 @@
 name = "AstroImages"
 uuid = "fe3fc30c-9b16-11e9-1c73-17dabf39f4ad"
 authors = ["Mosè Giordano", "Rohit Kumar", "William Thompson"]
-version = "0.4.0"
+version = "0.4.1"
 
 [deps]
 AbstractFFTs = "621f4979-c628-5d54-868e-fcf4e3e8185c"
@@ -28,7 +28,7 @@ WCS = "15f3aee2-9e10-537f-b834-a6fb8bdb944d"
 AbstractFFTs = "1.1"
 AstroAngles = "0.1"
 ColorSchemes = "3.18"
-DimensionalData = "0.23"
+DimensionalData = "0.23, 0.24"
 FITSIO = "0.16, 0.17"
 FileIO = "1.15"
 ImageAxes = "0.6"
