### 2023-03-20 #126 @whedon
[REVIEW]: AstroImages.jl
**Submitting author:** @sefffal (<a href="http://orcid.org/0000-0001-5684-4593">William R. Thompson</a>)
**Repository:** https://github.com/JuliaAstro/AstroImages.jl.git
**Branch with paper.md** (empty if default branch): paper2
**Version:**

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?

- [ ] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
