### 2023-03-20 #126 @carstenbauer
@oschulz, @lucaferranti
**Please take a look at the paper here:** https://github.com/JuliaAstro/AstroImages.jl/blob/paper/paper/paper.pdf
It would be great to get your reviews in within the next two weeks. Please let me know if you need more time. Thanks in advance!
