### 2023-03-21 #126 @lucaferranti
I can start reviewing the software side, but for the paper, I'd prefer to be able to compile the pdf on demand with wheedon and have the latest version available. The manually added pdf gets easily outdated
