### 2023-03-21 #126 @sefffal
Hi all, thanks for continuing with the review. It’s still not clear to me why the pdf won’t compile with wheedon and will need outside help to fix the issue.

The author list consists of myself and @giordano. The metadata file should be correct so hopefully when the above issue is solved the author list will be inserted correctly into the PDF.
