### 2023-04-25 afd2118
Point docs link to JuliaAstro general docs page
