### 2023-05-05 #126 @sefffal
Hmm strange, in order to debug the failure I added a post build command to latexmkrc to echo `paper.log`. And now it succeeded in generating a proof?
