+++ paper/header.tex
@@ -1,18 +1,17 @@
 % **************GENERATED FILE, DO NOT EDIT**************
+% I editted this manually anyways since whedon isn't updating it? IDK.
+\title{AstroImages.jl}
 
-\title{My JuliaCon proceeding}
+\author[1]{William Thompson}
+\author[2]{Mosè Giordano}
+\affil[1]{University of Victoria}
+\affil[2]{Advanced Research Computing, University College London}
 
-\author[1]{1st author}
-\author[1, 2]{2nd author}
-\author[2]{3rd author}
-\affil[1]{University}
-\affil[2]{National Lab}
-
-\keywords{Julia, Optimization, Game theory, Compiler}
+\keywords{Julia, Astronomy, Images, World Coordinates}
 
 \hypersetup{
 pdftitle = {My JuliaCon proceeding},
 pdfsubject = {JuliaCon 2019 Proceedings},
-pdfauthor = {1st author, 2nd author, 3rd author},
+pdfauthor = {William Thompson, Mosè Giordano},
 pdfkeywords = {Julia, Optimization, Game theory, Compiler},
 }
