+++ paper/paper.yml
@@ -8,7 +8,7 @@ authors:
   - name: William Thompson
     orcid: 0000-0001-5684-4593
     affiliation: 1
-  - name: Mosè Giordano
+  - name: Mose Giordano
     orcid: 0000-0002-7218-2873
     affiliation: 2
 affiliations:
