+++ paper/paper.tex
@@ -5,8 +5,23 @@
 
 \begin{document}
 
-\input{header}
+% \input{header}
 
+\title{AstroImages.jl}
+
+\author[1]{William Thompson}
+\author[2]{Mosè Giordano}
+\affil[1]{University of Victoria}
+\affil[2]{Advanced Research Computing, University College London}
+
+\keywords{Julia, Astronomy, Images, World Coordinates}
+
+\hypersetup{
+pdftitle = {AstroImages.jl},
+pdfsubject = {JuliaCon 2019 Proceedings},
+pdfauthor = {William Thompson, Mosè Giordano},
+pdfkeywords = {Julia, Optimization, Game theory, Compiler},
+}
 
 \title{AstroImages.jl}
 
