+++ paper/paper.tex
@@ -7,7 +7,7 @@
 
 % \input{header}
 
-\title{AstroImages.jl}
+\title{AstroImages.jl Test}
 
 \author[1]{William Thompson}
 \author[2]{Mosè Giordano}
@@ -23,7 +23,6 @@ pdfauthor = {William Thompson, Mosè Giordano},
 pdfkeywords = {Julia, Optimization, Game theory, Compiler},
 }
 
-\title{AstroImages.jl}
 
 \maketitle
 
