+++ paper/paper.tex
@@ -7,25 +7,26 @@
 
 % \input{header}
 
-\title{AstroImages.jl Test}
+% \title{AstroImages.jl Test}
 
-\author[1]{William Thompson}
-\author[2]{Mosè Giordano}
-\affil[1]{University of Victoria}
-\affil[2]{Advanced Research Computing, University College London}
+% \author[1]{William Thompson}
+% \author[2]{Mosè Giordano}
+% \affil[1]{University of Victoria}
+% \affil[2]{Advanced Research Computing, University College London}
 
-\keywords{Julia, Astronomy, Images, World Coordinates}
+% \keywords{Julia, Astronomy, Images, World Coordinates}
 
-\hypersetup{
-pdftitle = {AstroImages.jl},
-pdfsubject = {JuliaCon 2019 Proceedings},
-pdfauthor = {William Thompson, Mosè Giordano},
-pdfkeywords = {Julia, Optimization, Game theory, Compiler},
-}
+% \hypersetup{
+% pdftitle = {AstroImages.jl},
+% pdfsubject = {JuliaCon 2019 Proceedings},
+% pdfauthor = {William Thompson, Mosè Giordano},
+% pdfkeywords = {Julia, Optimization, Game theory, Compiler},
+% }
 
 
-\maketitle
+% \maketitle
 
+TEST TEST TEST
 To study the cosmos, astronomers use data cubes with many dimensions representing images with axes for sky position, time, wavelength, polarization, and more. Since these large datasets often span many orders of magnitude in intensity and typically include colours invisible to humans, astronomers like to visualize their images using a variety of non-linear stretching and contrast adjustments.
 Additionally, images may contain metadata specifying arbitrary mappings of pixel positions to multiple celestial coordinate systems.
 Julia\cite{bezanson2017julia} is a powerful language for processing astronomical data, but these visualization tasks are a challenge for any tool.
