### 2023-05-05 #126 @sefffal
@oschulz @carstenbauer I can confirm that the PDF generated by whedon does not reflect the contents of the repository.
I even tried manually inserting the authors and title instead of letting whedon insert it with the ruby script.
Changing the text also has no effect.

I tried to echo the latex log but I'm not sure where that goes.

Unfortunately I am now truly out of ideas on how to debug this without digging into the whedon infrastructure itself.
