### 2023-05-24 8242889
Correct imview function docstring
