### 2023-06-20 #126 @lucaferranti
apologies for the long radio silence, I've now finally managed to review from the software side and I have only few minor nitpicks

1. The pages `Polarization` and `Spectral Axes` in the manual are empty
2. While the organization docs has a short `Contributing` section in the homepage, for the purpose of submission it would be good to add a short minimalistic contributing guidelines also for the package readme.

3. (this is an extra improvement suggestion, not needed for this submission). One thing I noticed while trying the examples in the docs is that it was downloading manually some images. Have you considered doing something like [TestImages.jl](https://github.com/JuliaImages/TestImages.jl) but for astro images? (maybe call it `AstroTestImages.jl`) That could offer a nicer interface for users to play with popular(?) astro images e.g. while going through the documentation.
