### 2023-06-20 #126 @lucaferranti
@sefffal should I generate the pdf from the branch `paper` or `paper2`? Can you take a quick look at the latest generated pdf and confirm whether from a content perspective it's up to date and ready for review?
