### 2023-06-20 #126 @sefffal
> @sefffal should I generate the pdf from the branch `paper` or `paper2`? Can you take a quick look at the latest generated pdf and confirm whether from a content perspective it's up to date and ready for review?

Excellent, thank you for your review! The paper proof is up to date content wise but is still missing the author and journal info. I tried my best to find the issue but for some reason whedon is not updating the linked PDF.
