### 2023-09-05 #126 @lucaferranti
@sefffal , the editorial team got some fresh blood and we are among other things upgrading the bot from the old `whedon` to `editorial-bot` which should make latex workflow easier. This should happen within the end of the month, but I can nevertheless also look into the specific issue of this and get this fixed before.

Meanwhile, I've actually almost finished the review from paper side and will post the comments soon (only minor comments), so this should go forward, apologies for the loong delay
