### 2023-09-07 #126 @lucaferranti
@sefffal, at least compilation runs without errors, but there's still the issue of authors not being inserted.

As a quick check, may I ask you to try the following?

1. Save the content of your paper, metadata in `paper.yml` and references in `ref.bib`
2. Overwrite the `paper` folder with a fresh template from https://github.com/JuliaCon/JuliaConSubmission.jl
3. Replace the new `paper.yml` and `ref.bib` with your own and reinser the paper body into `paper.tex`

If this work, then we can all be happy. If it doesnt, then I'll try to narrow down the issue further.
