+++ paper/header.tex
@@ -1,17 +1,19 @@
 % **************GENERATED FILE, DO NOT EDIT**************
-% I editted this manually anyways since whedon isn't updating it? IDK.
-\title{AstroImages.jl}
 
-\author[1]{William Thompson}
-\author[2]{Mosè Giordano}
-\affil[1]{University of Victoria}
-\affil[2]{Advanced Research Computing, University College London}
+\title{My JuliaCon proceeding}
 
-\keywords{Julia, Astronomy, Images, World Coordinates}
+\author[1]{1st author}
+\author[1, 2]{2nd author}
+\author[2]{3rd author}
+\affil[1]{University}
+\affil[2]{National Lab}
+
+\keywords{Julia, Optimization, Game theory, Compiler}
 
 \hypersetup{
 pdftitle = {My JuliaCon proceeding},
-pdfsubject = {JuliaCon 2019 Proceedings},
-pdfauthor = {William Thompson, Mosè Giordano},
+pdfsubject = {JuliaCon 2022 Proceedings},
+pdfauthor = {1st author, 2nd author, 3rd author},
 pdfkeywords = {Julia, Optimization, Game theory, Compiler},
 }
+
