+++ paper/paper.yml
@@ -1,22 +1,21 @@
-title: "AstroImages.jl"
+title: "My JuliaCon proceeding test"
 keywords:
   - Julia
-  - Astronomy
-  - Images
-  - World Coordinates
+  - Optimization
+  - Game theory
+  - Compiler
 authors:
   - name: William Thompson
-    orcid: 0000-0001-5684-4593
+    orcid: 0000-0003-1361-440X
     affiliation: 1
-  - name: Mose Giordano
-    orcid: 0000-0002-7218-2873
+  - name: 2nd author
+    affiliation: 1, 2
+  - name: 3rd author
     affiliation: 2
 affiliations:
-  - name: University of Victoria
+  - name: University
     index: 1
-  - name: Advanced Research Computing, University College London
+  - name: National Lab
     index: 2
-date: 02 August 2022
+date: 24 April 2019
 bibliography: ref.bib
-
-
