### 2023-09-07 #126 @sefffal
As a first test, I have reset the the paper folder (paper branch) to the exact contents of the template. I'll try to build this and see if it works.

If that works, I'll add back in our paper content.
