+++ paper/paper.tex
@@ -9,532 +9,122 @@
 
 \maketitle
 
-\begin{abstract}
+To study the cosmos, astronomers use data cubes with many dimensions representing images with axes for sky position, time, wavelength, polarization, and more. Since these large datasets often span many orders of magnitude in intensity and typically include colours invisible to humans, astronomers like to visualize their images using a variety of non-linear stretching and contrast adjustments.
+Additionally, images may contain metadata specifying arbitrary mappings of pixel positions to multiple celestial coordinate systems.
+Julia\cite{bezanson2017julia} is a powerful language for processing astronomical data, but these visualization tasks are a challenge for any tool.
 
-This is a guide for authors who are preparing papers for JuliaCon using the \LaTeX{} document
-preparation system and the \verb|juliacon|  class file.
+\vspace{0.2cm}
+\emph{Background.}
+One of the most ubiquitous data formats used in astronomy is FITS, or the Flexible Image Transport System.
+Compared to a traditional raster image formats, FITS files are more like containers or small filesystems.
+Each file contains one or more header data units (HDUs) that pair a dataset with a header.
+The FITS format and header conventions have developed organically over several decades, first by convention,
+and then in a series of papers proposing new extensions and standardizing existing behaviour now summarized in \cite{fitsstandard}.
 
-\end{abstract}
+In each HDU,
+headers are stored in a plain text ASCII format and are followed by data.
+The data may be in the format of an N-dimensional binary array (an ``image'' HDU),
+a binary table,
+or an ASCII table (``table HDUs'').
 
-\section{Introduction}
-
-The \LaTeX{} document preparation system is a special version of the
-\TeX; typesetting program where-in a collection of \TeX; commands
-are added to \LaTeX{} to simplify typesetting. Importantly, it allows
-the author to concentrate on the logical structure of the document
-rather than its visual layout.\vskip 6pt
-Moreover, \LaTeX{} provides a consistent and comprehensive document
-preparation interface. There are simple-to-use commands for
-generating a table of contents, lists of figures and/or tables, and indexes.
-\LaTeX{} can automatically number list entries, equations, figures,
-tables, and footnotes, as well as articles, sections, and subsections.
-Using this numbering system, bibliographic citations, page
-references, and cross references to any other numbered entity (e.g.
-article, section, equation, figure, list entry, etc.) become quite simple
-and straightforward. The use of \LaTeX{} document classes allows
-a simple change of class to transform the appearance of your document.\vskip 6pt
-\LaTeX{} is a powerful tool for managing long and complex documents.
-In particular, partial processing enables long documents to
-be produced article by article without losing sequential information.
-The use of document classes allows a simple change of style
-(or style option) to transform the appearance of your document.
+The metadata described by FITS headers are quite rich.
+A header consists of \verb|KEY = VALUE / COMMENT| entries, long form \verb|COMMENT| sections,
+and long form \verb|HISTORY| sections used to desrcibe the sequence of transformations used to generate the data.
+FITS headers often contain several hundred header entries most of which are specific to the instrument and software packages used to record and process the data.
+A subset of these are standardized and used to describe the physical units and coordinates of pixels in the image, called world coordinate system (WCS) header entries.
 
 
-
-\section{The JuliaCon Article Class}
[…]
