### 2023-09-07 77ebc64
Diagnose paper compilation issues
