+++ paper/paper.tex
@@ -119,7 +119,7 @@ implot(slice; cmap=:turbo, stretch=asinh)
 \vspace{-0.5cm}
 \begin{figure}[hbt!]
     \centerline{\includegraphics[width=\columnwidth]{hidatplot.pdf}}
-    \caption{implot augments an image with world coordinates and automatically labelled axes, titles, and colourbar. Image is a slice through galactic latitude, longitude, and velocity space of neutral Hydrogen density (HI4PI survey \cite{2016A&A...594A.116H}).}
+    \caption{implot augments an image with world coordinates and automatically labelled axes, titles, and colourbar. Image is a slice through galactic latitude, longitude, and velocity space of neutral Hydrogen density (HI4PI survey \cite{HI4PI}).}
       \label{fig:cube-slice}
 \end{figure}
 \FloatBarrier 
