### 2023-09-07 #126 @sefffal
So after these tests, there's no issues with `paper.yml`. The issues start after I swapped out the body of the paper template and `ref.bib`.

Building locally, the references are picked up as intended.
