+++ paper/paper.tex
@@ -11,7 +11,7 @@
 
 To study the cosmos, astronomers use data cubes with many dimensions representing images with axes for sky position, time, wavelength, polarization, and more. Since these large datasets often span many orders of magnitude in intensity and typically include colours invisible to humans, astronomers like to visualize their images using a variety of non-linear stretching and contrast adjustments.
 Additionally, images may contain metadata specifying arbitrary mappings of pixel positions to multiple celestial coordinate systems.
-Julia\cite{bezanson2017julia} is a powerful language for processing astronomical data, but these visualization tasks are a challenge for any tool.
+Julia is a powerful language for processing astronomical data, but these visualization tasks are a challenge for any tool.
 
 \vspace{0.2cm}
 \emph{Background.}
@@ -19,7 +19,7 @@ One of the most ubiquitous data formats used in astronomy is FITS, or the Flexib
 Compared to a traditional raster image formats, FITS files are more like containers or small filesystems.
 Each file contains one or more header data units (HDUs) that pair a dataset with a header.
 The FITS format and header conventions have developed organically over several decades, first by convention,
-and then in a series of papers proposing new extensions and standardizing existing behaviour now summarized in \cite{fitsstandard}.
+and then in a series of papers proposing new extensions and standardizing existing behaviour now summarized in.
 
 In each HDU,
 headers are stored in a plain text ASCII format and are followed by data.
@@ -119,7 +119,7 @@ implot(slice; cmap=:turbo, stretch=asinh)
 \vspace{-0.5cm}
 \begin{figure}[hbt!]
     \centerline{\includegraphics[width=\columnwidth]{hidatplot.pdf}}
-    \caption{implot augments an image with world coordinates and automatically labelled axes, titles, and colourbar. Image is a slice through galactic latitude, longitude, and velocity space of neutral Hydrogen density (HI4PI survey \cite{HI4PI}).}
+    \caption{implot augments an image with world coordinates and automatically labelled axes, titles, and colourbar. Image is a slice through galactic latitude, longitude, and velocity space of neutral Hydrogen density (HI4PI survey).}
       \label{fig:cube-slice}
 \end{figure}
 \FloatBarrier 
