### 2023-09-07 #126 @sefffal
For this test above, I removed all uses of `\cite`. Looks like that wasn't the issue, and there is something deeper. I will fully revert paper.tex and add paragraphs one by one...
