+++ paper/paper.tex
@@ -9,9 +9,16 @@
 
 \maketitle
 
+\begin{abstract}
+
+This is a guide for authors who are preparing papers for JuliaCon using the \LaTeX{} document
+preparation system and the \verb|juliacon|  class file.
+
+\end{abstract}
+
 To study the cosmos, astronomers use data cubes with many dimensions representing images with axes for sky position, time, wavelength, polarization, and more. Since these large datasets often span many orders of magnitude in intensity and typically include colours invisible to humans, astronomers like to visualize their images using a variety of non-linear stretching and contrast adjustments.
 Additionally, images may contain metadata specifying arbitrary mappings of pixel positions to multiple celestial coordinate systems.
-Julia is a powerful language for processing astronomical data, but these visualization tasks are a challenge for any tool.
+Julia\cite{bezanson2017julia} is a powerful language for processing astronomical data, but these visualization tasks are a challenge for any tool.
 
 \vspace{0.2cm}
 \emph{Background.}
@@ -19,7 +26,7 @@ One of the most ubiquitous data formats used in astronomy is FITS, or the Flexib
 Compared to a traditional raster image formats, FITS files are more like containers or small filesystems.
 Each file contains one or more header data units (HDUs) that pair a dataset with a header.
 The FITS format and header conventions have developed organically over several decades, first by convention,
-and then in a series of papers proposing new extensions and standardizing existing behaviour now summarized in.
+and then in a series of papers proposing new extensions and standardizing existing behaviour now summarized in \cite{fitsstandard}.
 
 In each HDU,
 headers are stored in a plain text ASCII format and are followed by data.
@@ -39,92 +46,525 @@ For most coordinate schemes, pixel coordinates are encoded using an affine trans
 In many instances, image HDUs contain 3, 4, or even 5 dimensional data cubes, and thanks to the affine transformation matrix, moving along any one of these dimensions can shift the coordinates along the other dimesions. This point is important since it means that the full coordinates of a given pixel are needed to calculate it's physical position along any one axis.
 
 
+\section{Introduction}
+
+The \LaTeX{} document preparation system is a special version of the
+\TeX; typesetting program where-in a collection of \TeX; commands
+are added to \LaTeX{} to simplify typesetting. Importantly, it allows
+the author to concentrate on the logical structure of the document
+rather than its visual layout.\vskip 6pt
+Moreover, \LaTeX{} provides a consistent and comprehensive document
+preparation interface. There are simple-to-use commands for
+generating a table of contents, lists of figures and/or tables, and indexes.
+\LaTeX{} can automatically number list entries, equations, figures,
+tables, and footnotes, as well as articles, sections, and subsections.
+Using this numbering system, bibliographic citations, page
+references, and cross references to any other numbered entity (e.g.
+article, section, equation, figure, list entry, etc.) become quite simple
+and straightforward. The use of \LaTeX{} document classes allows
+a simple change of class to transform the appearance of your document.\vskip 6pt
+\LaTeX{} is a powerful tool for managing long and complex documents.
+In particular, partial processing enables long documents to
+be produced article by article without losing sequential information.
+The use of document classes allows a simple change of style
+(or style option) to transform the appearance of your document.
+
+
+
[…]
