### 2023-09-07 e003756
Test removing the figure
