+++ paper-old/paper.tex
@@ -1,0 +1,153 @@
+
+% JuliaCon proceedings template
+\documentclass{juliacon}
+\setcounter{page}{1}
+
+\begin{document}
+
+% \input{header}
+
+\title{AstroImages.jl}
+
+\author[1]{William Thompson}
+\author[2]{Mosè Giordano}
+\affil[1]{University of Victoria}
+\affil[2]{Advanced Research Computing, University College London}
+
+\keywords{Julia, Astronomy, Images, World Coordinates}
+
+\hypersetup{
+pdftitle = {AstroImages.jl},
+pdfsubject = {JuliaCon 2019 Proceedings},
+pdfauthor = {William Thompson, Mosè Giordano},
+pdfkeywords = {Julia, Optimization, Game theory, Compiler},
+}
+
+\title{AstroImages.jl}
+
+\maketitle
+
+To study the cosmos, astronomers use data cubes with many dimensions representing images with axes for sky position, time, wavelength, polarization, and more. Since these large datasets often span many orders of magnitude in intensity and typically include colours invisible to humans, astronomers like to visualize their images using a variety of non-linear stretching and contrast adjustments.
+Additionally, images may contain metadata specifying arbitrary mappings of pixel positions to multiple celestial coordinate systems.
+Julia\cite{bezanson2017julia} is a powerful language for processing astronomical data, but these visualization tasks are a challenge for any tool.
+
+\vspace{0.2cm}
+\emph{Background.}
+One of the most ubiquitous data formats used in astronomy is FITS, or the Flexible Image Transport System.
+Compared to a traditional raster image formats, FITS files are more like containers or small filesystems.
+Each file contains one or more header data units (HDUs) that pair a dataset with a header.
+The FITS format and header conventions have developed organically over several decades, first by convention,
+and then in a series of papers proposing new extensions and standardizing existing behaviour now summarized in \cite{fitsstandard}.
+
+In each HDU,
+headers are stored in a plain text ASCII format and are followed by data.
+The data may be in the format of an N-dimensional binary array (an ``image'' HDU),
+a binary table,
+or an ASCII table (``table HDUs'').
+
+The metadata described by FITS headers are quite rich.
+A header consists of \verb|KEY = VALUE / COMMENT| entries, long form \verb|COMMENT| sections,
+and long form \verb|HISTORY| sections used to desrcibe the sequence of transformations used to generate the data.
+FITS headers often contain several hundred header entries most of which are specific to the instrument and software packages used to record and process the data.
+A subset of these are standardized and used to describe the physical units and coordinates of pixels in the image, called world coordinate system (WCS) header entries.
+
+
+These WCS entries record pixel locations in one or more celestial coordinate systems including, right ascension and declination, galactic coordinates, velocity, frequency, wavelength, polarization, and more.
+For most coordinate schemes, pixel coordinates are encoded using an affine transformation matrix combined with a \verb|CTYPE| specifier. Images are typically sampled regularly in the plane of a detector; however, coordinate projections are in general non-linear, and the physical coordinates of each pixel, spacings between pixels, etc. may vary across the image.
+In many instances, image HDUs contain 3, 4, or even 5 dimensional data cubes, and thanks to the affine transformation matrix, moving along any one of these dimensions can shift the coordinates along the other dimesions. This point is important since it means that the full coordinates of a given pixel are needed to calculate it's physical position along any one axis.
+
+
+\begin{figure}[t]
+    \centerline{\includegraphics[width=\columnwidth]{fits-schematic.pdf}}
[…]
