+++ paper/paper.tex
@@ -46,11 +46,6 @@ For most coordinate schemes, pixel coordinates are encoded using an affine trans
 In many instances, image HDUs contain 3, 4, or even 5 dimensional data cubes, and thanks to the affine transformation matrix, moving along any one of these dimensions can shift the coordinates along the other dimesions. This point is important since it means that the full coordinates of a given pixel are needed to calculate it's physical position along any one axis.
 
 
-\begin{figure}[t]
-  \centerline{\includegraphics[width=\columnwidth]{fits-schematic.pdf}}
-  \caption{Conceptual schematic of a FITS file containing a 2D image HDU, a 3D hyperspectral image HDU, and a table HDU.}
-      \label{fig:fits-schematic}
-\end{figure}
 
 For image data itself, FITS files do not in general store colour information that can be displayed directly. In contrast to digital cameras, astronomical data is almost invariably captured outside the human visible range or at least with filters that are not well matched to human colour perception.
 Instead, astronomers use visualizations that maps raw numerical data to a false colour image.
