+++ docs/src/AstroImagesdocs.jl
@@ -1,0 +1,1 @@
+module AstroImagesdocs; end
