+++ paper-old/ref.bib
@@ -1,0 +1,40 @@
+@article{bezanson2017julia,
+  title={Julia: A fresh approach to numerical computing},
+  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  doi={10.1137/141000671},
+  journal={SIAM review},
+  volume={59},
+  number={1},
+  pages={65--98},
+  year={2017},
+  publisher={SIAM}
+}
+
+
+@article{2016A&A...594A.116H,
+  author        = {{HI4PI Collaboration} and {Ben Bekhti}, N. and {Fl{\"o}er}, L. and {Keller}, R. and {Kerp}, J. and {Lenz}, D. and {Winkel}, B. and {Bailin}, J. and {Calabretta}, M.~R. and {Dedes}, L. and {Ford}, H.~A. and {Gibson}, B.~K. and {Haud}, U. and {Janowiecki}, S. and {Kalberla}, P.~M.~W. and {Lockman}, F.~J. and {McClure-Griffiths}, N.~M. and {Murphy}, T. and {Nakanishi}, H. and {Pisano}, D.~J. and {Staveley-Smith}, L.},
+  title         = {{HI4PI: A full-sky H I survey based on EBHIS and GASS}},
+  journal       = {\aap},
+  keywords      = {surveys, ISM: atoms, techniques: spectroscopic, Astrophysics - Astrophysics of Galaxies},
+  year          = 2016,
+  month         = oct,
+  volume        = {594},
+  eid           = {A116},
+  pages         = {A116},
+  doi           = {10.1051/0004-6361/201629178},
+  archiveprefix = {arXiv},
+  eprint        = {1610.06175},
+  primaryclass  = {astro-ph.GA},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2016A&A...594A.116H},
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@techreport{fitsstandard,
+  title       = {Definition of the Flexible Image Transport System
+{(FITS)}},
+  author      = {{FITS Working Group}},
+  institution = {International Astronomical Union},
+  address     = {},
+  number      = {},
+  year        = 2016,
+}
