+++ paper-old/header.tex
@@ -1,0 +1,17 @@
+% **************GENERATED FILE, DO NOT EDIT**************
+% I editted this manually anyways since whedon isn't updating it? IDK.
+\title{AstroImages.jl}
+
+\author[1]{William Thompson}
+\author[2]{Mosè Giordano}
+\affil[1]{University of Victoria}
+\affil[2]{Advanced Research Computing, University College London}
+
+\keywords{Julia, Astronomy, Images, World Coordinates}
+
+\hypersetup{
+pdftitle = {My JuliaCon proceeding},
+pdfsubject = {JuliaCon 2019 Proceedings},
+pdfauthor = {William Thompson, Mosè Giordano},
+pdfkeywords = {Julia, Optimization, Game theory, Compiler},
+}
