+++ paper-old/paper.yml
@@ -1,0 +1,22 @@
+title: "AstroImages.jl"
+keywords:
+  - Julia
+  - Astronomy
+  - Images
+  - World Coordinates
+authors:
+  - name: William Thompson
+    orcid: 0000-0001-5684-4593
+    affiliation: 1
+  - name: Mose Giordano
+    orcid: 0000-0002-7218-2873
+    affiliation: 2
+affiliations:
+  - name: University of Victoria
+    index: 1
+  - name: Advanced Research Computing, University College London
+    index: 2
+date: 02 August 2022
+bibliography: ref.bib
+
+
