### 2023-10-25 #126 @lucaferranti
@arfon @xuanxu could you help figuring out why whedon does not produce the pdf (nor informs about an error) are the logs of the action available somewhere?
