### 2023-10-25 #126 @xuanxu
The problem is caused by the presence of more than one paper.md/.tex file
