### 2023-10-25 #126 @pitsianis
@xuanxu thank you! It would be good the `generate pdf from` command to provide a link to the compilation (error) logs, even though the link would get invalidated after a little while so that we do not have to bother you that often :-) Alternatively, provide the compilation environment and scripts so a motivated user can replicate the `latexmk` command independently.

@sefffal please make the suggested correction.
