### 2023-10-30 #126 @sefffal
Thanks all for helping to identify the problem. Could you clarify it a bit more for me though, in branch `paper` I have a folder called `paper` with `paper.tex`. The only other `.tex` files come from the template, and there are no `.md` files.
