### 2024-03-23 #126 @lucaferranti
@sefffal pdf generation works :tada: . I wasn't sure which branch to use `paper` or `paper2`, can you take a look at the paper and let me know?
